+++ main/src/main/java/org/sopt/makers/crew/main/external/auth/dto/request/PlaygroundUserRequestDto.java
@@ -1,0 +1,9 @@
+package org.sopt.makers.crew.main.external.auth.dto.request;
+
+public record PlaygroundUserRequestDto(
+	String accessToken
+) {
+	public static PlaygroundUserRequestDto from(String accessToken) {
+		return new PlaygroundUserRequestDto(accessToken);
+	}
+}
