--- main/src/main/java/org/sopt/makers/crew/main/external/playground/dto/request/PlaygroundUserRequestDto.java
@@ -1,11 +1,0 @@
-package org.sopt.makers.crew.main.external.playground.dto.request;
-
-
-public record PlaygroundUserRequestDto(
-	String accessToken
-
-) {
-	public static PlaygroundUserRequestDto of(String accessToken){
-		return new PlaygroundUserRequestDto(accessToken);
-	}
-}
