+++ main/src/main/java/org/sopt/makers/crew/main/global/jwt/JwtTokenProvider.java
@@ -6,6 +6,7 @@ import java.util.Date;
 
 import javax.crypto.spec.SecretKeySpec;
 
+import org.sopt.makers.crew.main.global.security.JwtExceptionType;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
